Speeds up extract_stars by a factor of 10
The center pixel of the star is used as an approximation of the
brightest pixel. This action much quicker than finding the actual brightest
pixel by generating the star histogram and then looping through to find
the highest non-zero intensity in the histogram array. This also cuts out
the overhead from masking the image which was a sizeable portion of
the original computation time.
Also fixes the placement of the docstring so the docstring will show
when calling extract_stars.__doc__